splitting out some code from game into other files
